Updated Hamburger Clicking
Fixed some unexpected functionality of the hamburger menu that appears when the page is narrow. The menu was not appearing which was fixed by flipping the logic of get-menu-state.js. Then the menu was opening on a click anywhere and not closing unless a link was clicked, which was fixed by updating navbar-container.js.